refactor: issue#11 Separate Message class from Slack class
close: #11

The logic that creates the notification message has nothing to do with Slack notifications,
so separate it from the Slack class.
